26/04/2022
-lul fixed it

Twitch being muted was actually a much bigger problem than I thought, but also my implementation was really bad (I had a webview of a custom hosted webpage that embeded the twitch player in it) and the twitch player obeys all the rules of a default browser (like chrome or etc) and those browsers have very strict rules about loading a video and playing it with sound without user interaction.

But like I said it was the wrong way to go about it anyway, because a webview of the link (youtube or twitch) doesn't involve a 3rd party browser (just the android WebView browser), and therefore I can custom set the flag which allows videos in a twitch or Youtube webpage to play automatically (mediaPlaybackRequiresUserGesture). There was a little fenangling to ensure the webview didn't by default use the native twitch or youtube mobile apps...? (I don't actually know what these are, they're not the downloaded apps, theyre like if you view a webpage from mobile, because what I want is just like the desktop site ((I think)))

Either way, links are playing on lockedscreen on alarm going off now. (Everything looks shit because of a 3-day proof that it could work worrying about nothing else)